requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0

# Academic Paper Search
arxiv>=2.1.0
semanticscholar>=0.8.0

# PDF Processing
pymupdf>=1.24.0
pdfplumber>=0.11.0

# Embeddings & Vector DB
sentence-transformers>=3.0.0
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Data & Utilities
aiohttp>=3.9.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0
httpx>=0.27.0

# UI
streamlit>=1.38.0
plotly>=5.18.0
streamlit-agraph>=0.0.45
networkx>=3.2.0

# Presentation / Output
python-pptx>=1.0.0
markdown>=3.5.0
jinja2>=3.1.0

# Logging & Monitoring
rich>=13.7.0
loguru>=0.7.0
